+++ src/App.jsx
@@ -1,7 +1,7 @@
 import { Layout } from './pages/Layout';
 import { LandingPage } from './pages/LandingPage';
 import { ListPage } from './pages/ListPage';
-import { PostPage } from './pages/PostCardPage';
+import { PostCardPage } from './pages/PostCardPage';
 import { PostMessagePage } from './pages/PostMessagePage';
 import { CardPostListPage } from './pages/CardPostListPage';
 import { BrowserRouter, Routes, Route } from 'react-router-dom';
@@ -13,7 +13,7 @@ function App() {
         <Route path="/" element={<Layout />}>
           <Route index element={<LandingPage />} />
           <Route path="list" element={<ListPage />} />
-          <Route path="post" element={<PostPage />} />
+          <Route path="post" element={<PostCardPage />} />
           <Route
             path="post/:recipientId/messages"
             element={<PostMessagePage />}
