+++ src/pages/PostCardPage/PostCardPage.module.scss
