+++ src/pages/PostCardPage/index.js
@@ -1,1 +1,1 @@
-export { default as PostPage } from './PostPage';
+export { default as PostCardPage } from './PostCardPage';
